TS for the component template
