--- helpers/.component-template/cypress.json
@@ -1,4 +1,0 @@
-{
-	"fileServerFolder": "examples/",
-	"integrationFolder": "tests/integration/"
-}
