--- helpers/.component-template/src/_COMPONENT_NAME_.js
@@ -1,32 +1,0 @@
-import { jsx } from '@westpac/core';
-import PropTypes from 'prop-types';
-
-// ==============================
-// Utils
-// ==============================
-
-// ==============================
-// Component
-// ==============================
-
-/**
- * _COMPONENT_TITLE_: _COMPONENT_DESCRIPTION_
- */
-export const _COMPONENT_NAME_ = ({ ...props }) => {
-	return (
-		<div css={{}} {...props}>
-			_COMPONENT_NAME_
-		</div>
-	);
-};
-
-// ==============================
-// Types
-// ==============================
-
-_COMPONENT_NAME_.propTypes = {
-	/**
-	 * Describe `someProperty` here
-	 */
-	someProperty: PropTypes.string,
-};
