+++ helpers/.component-template/src/_COMPONENT_NAME_.tsx
@@ -1,0 +1,15 @@
+import PropTypes from 'prop-types';
+
+/** _COMPONENT_TITLE_: _COMPONENT_DESCRIPTION_ */
+export const _COMPONENT_NAME_ = ({ ...props }) => {
+	return (
+		<div css={{}} {...props}>
+			_COMPONENT_NAME_
+		</div>
+	);
+};
+
+_COMPONENT_NAME_.propTypes = {
+	/** * Describe `someProperty` here */
+	someProperty: PropTypes.string,
+};
