--- helpers/.component-template/tests/integration/_COMPONENT_KEY_.cypress.js
@@ -1,5 +1,0 @@
-describe('_COMPONENT_TITLE_', () => {
-	before(() => {
-		cy.visit(`http://localhost:8080/`);
-	});
-});
