--- helpers/.component-template/tests/unit/_COMPONENT_KEY_.spec.js
@@ -1,3 +1,0 @@
-test('TODO replace this', () => {
-	expect(true).toBe(true);
-});
